+++ src/main.rs
@@ -1,5 +1,6 @@
 use atty::Stream;
 use clap::Parser;
+use colored::Colorize;
 
 mod api;
 mod repl;
@@ -31,15 +32,45 @@ pub struct Args {
     #[arg(short, long, help = "Enable verbose output for debugging")]
     pub verbose: bool,
 
-    #[arg(short, long, help = "Read commands from a file")]
+    #[arg(
+        short,
+        long,
+        help = "Read FSQL command from a file",
+        conflicts_with = "command"
+    )]
     pub file: Option<String>,
+
+    #[arg(
+        short,
+        long,
+        help = "Execute FSQL command directly",
+        conflicts_with = "file"
+    )]
+    pub command: Option<String>,
 }
 
 fn main() {
     let args = Args::parse();
 
-    // Check for file input first, then piped input, then REPL
-    if let Some(file_path) = args.file.clone() {
+    // Check for explicit input methods, then piped input, then REPL
+    if let Some(command) = args.command.clone() {
+        if !atty::is(Stream::Stdin) {
+            eprintln!(
+                "{}",
+                "❌ Cannot pipe to stdin and pass a command directly at the same time".red(),
+            );
+            std::process::exit(1);
+        }
+        let api_url = format!("https://{}/{}", args.host, args.path);
+        stdio::process_command(&command, &api_url, &args.token, args.verbose);
+    } else if let Some(file_path) = args.file.clone() {
+        if !atty::is(Stream::Stdin) {
+            eprintln!(
+                "{}",
+                "❌ Cannot pipe to stdin and pass a file at the same time".red(),
+            );
+            std::process::exit(1);
+        }
         stdio::handle_file(args, &file_path);
     } else if !atty::is(Stream::Stdin) {
         stdio::handle_stdin(args);